feat: backend completamente finalizado

+++ ERP-Backend/src/main/java/com/abbainc/erp/Repository/CamisaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CamisaRepository extends JpaRepository<Camisa, Long> {
+public interface CamisaRepository extends JpaRepository<Camisa, Integer> {
 
     Optional<Camisa> findBySku(String sku);
 }
